core: clk-stm32mp15: fix build error on enumarated
Fix build error reported by GCC 9.2 as shown in trace below:

core/drivers/clk/clk-stm32mp15.c:1090:2: error: case value ‘0’ not in enumerated type ‘enum stm32mp1_parent_id’ [-Werror=switch]
 1090 |  case _HSI:
      |  ^~~~
core/drivers/clk/clk-stm32mp15.c:1098:2: error: case value ‘1’ not in enumerated type ‘enum stm32mp1_parent_id’ [-Werror=switch]
 1098 |  case _HSE:
      |  ^~~~
core/drivers/clk/clk-stm32mp15.c:1094:2: error: case value ‘2’ not in enumerated type ‘enum stm32mp1_parent_id’ [-Werror=switch]
 1094 |  case _CSI:
      |  ^~~~
core/drivers/clk/clk-stm32mp15.c:1105:2: error: case value ‘3’ not in enumerated type ‘enum stm32mp1_parent_id’ [-Werror=switch]
 1105 |  case _LSI:
      |  ^~~~
core/drivers/clk/clk-stm32mp15.c:1108:2: error: case value ‘4’ not in enumerated type ‘enum stm32mp1_parent_id’ [-Werror=switch]
 1108 |  case _LSE:
      |  ^~~~
core/drivers/clk/clk-stm32mp15.c:1149:2: error: case value ‘6’ not in enumerated type ‘enum stm32mp1_parent_id’ [-Werror=switch]
 1149 |  case _USB_PHY_48:
      |  ^~~~

The issue comes from IDs used a parent clock identifiers come from 2
enumerated type: enum stm32mp_osc_id and enum stm32mp1_parent_id.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/194269
Tested-by: Etienne CARRIERE <etienne.carriere@foss.st.com>

+++ core/drivers/clk/clk-stm32mp15.c
@@ -962,7 +962,7 @@ static unsigned long __clk_get_parent_rate(enum stm32mp1_parent_id p)
 	unsigned long clock = 0;
 	vaddr_t rcc_base = stm32_rcc_base();
 
-	switch (p) {
+	switch ((unsigned int)p) {
 	case _CK_MPU:
 	/* MPU sub system */
 		reg = io_read32(rcc_base + RCC_MPCKSELR);
